requirements.txt
presidio-analyzer
presidio-anonymizer
python-dotenv
openai
